Avoid integer→float division artifacts and enforce SpotBugs ICAST_IDIV_CAST_TO_DOUBLE (#4406)
* Fix floating point divisions flagged by SpotBugs

* Avoid integer→float division artifacts and enforce SpotBugs ICAST_IDIV_CAST_TO_DOUBLE

### Motivation
- Address SpotBugs findings where an integral division result was cast to `double`/`float`, which can produce incorrect values.
- Keep rendering, geometry, and animation math semantically unchanged while using explicit floating-point operations to avoid integer truncation.
- Ensure the CI quality gate treats these SpotBugs findings as build failures by adding the rule to the forbidden list.

### Description
- Use explicit floating-point division/casts in chart and rendering code: `ChartComponent` (`zoomX`/`zoomY`), `RadarChart` (`centerX`/`centerY`), and `XYChart` (`mTranslate`).
- Fix UI/graphics math to use floats where appropriate: `CodenameOneImplementation.fillRectRadialGradient`, `CommonProgressAnimations.CircleProgress.stepSize`, `FlipTransition` (camera/translate math), `PerspectiveCamera` (camera math), `RoundRectBorder` (stroke offset), and `DefaultLookAndFeel` (state icon Y position).
- Fix numeric coefficient calculation in `Geometry.BezierCurve` by ensuring factorial division produces a `double`, and correct `StyleParser.getMMValue` to return a `float` with proper rounding/scale.
- Enforce SpotBugs rule `ICAST_IDIV_CAST_TO_DOUBLE` as a forbidden rule in `.github/scripts/generate-quality-report.py` so such findings fail the build.

### Testing
- No automated test suite was executed as part of this change.
- Static inspection and local code browsing were used to verify the numeric fixes compile and the SpotBugs rule was added to the quality-report script.

* Fix integer→float division artifacts and enforce SpotBugs ICAST_IDIV_CAST_TO_DOUBLE

### Motivation
- Prevent incorrect numeric results caused by integer division being cast to `float`/`double` in rendering and chart math.
- Silence SpotBugs `ICAST_IDIV_CAST_TO_DOUBLE` findings by fixing actual occurrences rather than masking them.
- Preserve the intended rendering/geometry semantics while ensuring explicit floating-point arithmetic where required.
- Improve precision in pinch-zoom, layout centering, camera transforms, and UI metrics.

### Description
- Replace integer division with explicit floating-point operations in chart and layout code such as `ChartComponent.pointerDragged`, `XYChart` center/translate, and `RadarChart` center calculations to avoid truncation.
- Adjust UI/graphics math in `DefaultLookAndFeel`, `RoundRectBorder`, `CodenameOneImplementation.fillRectRadialGradient`, `FlipTransition`, and `PerspectiveCamera` to use `0.5f`/`... * 0.5f` or explicit casts where appropriate.
- Fix numeric coefficient calculation by casting factorial division to `double` in `Geometry` and correct `StyleParser.getMMValue` to return a properly computed `float`; also refine `CommonProgressAnimations.CircleProgress` timing calculation.
- Add `ICAST_IDIV_CAST_TO_DOUBLE` to the forbidden SpotBugs rules in `.github/scripts/generate-quality-report.py` so such findings will fail the quality gate.

### Testing
- No unit or integration test suite was executed as part of this change.
- Static SpotBugs analysis was the motivating CI check and the edits target the previously reported `ICAST_IDIV_CAST_TO_DOUBLE` violations.
- The repository previously failed the SpotBugs quality gate due to these violations and the changes remove the identified integer→float division instances.
- CI re-run of the full pipeline was not performed within this patch set.

+++ CodenameOne/src/com/codename1/charts/ChartComponent.java
@@ -486,7 +486,7 @@ public class ChartComponent extends Component {
                 double[] panLimits = xyChart.getRenderer().getPanLimits();
 
                 if (zoomStart == null) {
-                    zoomStart = new Point((x[0] + x[1]) / 2, (y[0] + y[1]) / 2);
+                    zoomStart = new Point((x[0] + x[1]) / 2f, (y[0] + y[1]) / 2f);
 
                     zoomDistStartX = Math.abs(x[0] - x[1]);
                     zoomDistStartY = Math.abs(y[0] - y[1]);
@@ -503,8 +503,8 @@ public class ChartComponent extends Component {
                     if (dx == 0) dx = 1;
                     if (dy == 0) dy = 1;
 
-                    double zoomX = zoomDistStartX / dx;
-                    double zoomY = zoomDistStartY / dy;
+                    double zoomX = (double) zoomDistStartX / dx;
+                    double zoomY = (double) zoomDistStartY / dy;
 
 
                     BBox newBounds = zoomStartBBox.scaleScreenCoords((float) zoomX, (float) zoomY);
@@ -571,19 +571,19 @@ public class ChartComponent extends Component {
                 }
             } else {
                 if (zoomStart == null) {
-                    zoomStart = new Point((x[0] + x[1]) / 2, (y[0] + y[1]) / 2);
+                    zoomStart = new Point((x[0] + x[1]) / 2f, (y[0] + y[1]) / 2f);
                     zoomTransformStart = Transform.makeIdentity();
                     if (transform != null) {
                         zoomTransformStart.concatenate(transform);
                     }
-                    int dx = Math.abs(x[0] - x[1]) / 2;
-                    int dy = Math.abs(y[0] - y[1]) / 2;
+                    double dx = Math.abs(x[0] - x[1]) / 2f;
+                    double dy = Math.abs(y[0] - y[1]) / 2f;
                     zoomDistStart = Math.sqrt(dx * dx + dy * dy);
 
 
                 } else {
-                    int dx = Math.abs(x[0] - x[1]) / 2;
-                    int dy = Math.abs(y[0] - y[1]) / 2;
+                    double dx = Math.abs(x[0] - x[1]) / 2f;
+                    double dy = Math.abs(y[0] - y[1]) / 2f;
                     double zoomDist = Math.sqrt(dx * dx + dy * dy);
                     if (zoomDist == 0) {
                         zoomDist = 1;
